Speedup is mostly done; The BIG Split!
The drawing speedup is mostly done, but the glowy trails are not working
just yet.. :(  Collisions have some weirdness where they are not firing
all the time.
This also has The BIG Split!  After much muttering by other team members
over the joy of thousands of lines in one file, we now have a
fulsomeness of files!  And there was a jumping juggernaut of joy!